requirements.txt
tweepy>=4.14.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
networkx>=3.1
python-dateutil>=2.8.2
